fix: remove  nin and ne  operator usage in the SDK and the sample app (#2672)
* fix: remove  and  operator usage in the SDK and the sample app

* fix: remove console log

* fix: change console log to warn

* fix: add improvemnts

* fix: import of debouncefunc

* fix: import of debouncefunc

* fix: restructure queryMembers, queryUsers and ACItriggersettings

* fix: error bubbling for suggestions in auto complete input
