+++ package/src/contexts/messageInputContext/MessageInputContext.tsx
@@ -63,17 +63,19 @@ import {
   ImageUpload,
   UnknownType,
 } from '../../types/types';
-import { compressedImageURI } from '../../utils/compressImage';
-import { removeReservedFields } from '../../utils/removeReservedFields';
 import {
   ACITriggerSettings,
   ACITriggerSettingsParams,
+  TriggerSettings,
+} from '../../utils/ACITriggerSettings';
+import { compressedImageURI } from '../../utils/compressImage';
+import { removeReservedFields } from '../../utils/removeReservedFields';
+import {
   FileState,
   FileStateValue,
   generateRandomId,
   getFileNameFromPath,
   isBouncedMessage,
-  TriggerSettings,
 } from '../../utils/utils';
 import { useAttachmentPickerContext } from '../attachmentPickerContext/AttachmentPickerContext';
 import { ChannelContextValue, useChannelContext } from '../channelContext/ChannelContext';
@@ -1073,25 +1075,30 @@ export const MessageInputProvider = <
   };
 
   const getTriggerSettings = () => {
-    let triggerSettings: TriggerSettings<StreamChatGenerics> = {};
-    if (channel) {
-      if (value.autoCompleteTriggerSettings) {
-        triggerSettings = value.autoCompleteTriggerSettings({
-          channel,
-          client,
-          emojiSearchIndex: value.emojiSearchIndex,
-          onMentionSelectItem: onSelectItem,
-        });
-      } else {
-        triggerSettings = ACITriggerSettings<StreamChatGenerics>({
-          channel,
-          client,
-          emojiSearchIndex: value.emojiSearchIndex,
-          onMentionSelectItem: onSelectItem,
-        });
+    try {
+      let triggerSettings: TriggerSettings<StreamChatGenerics> = {};
+      if (channel) {
+        if (value.autoCompleteTriggerSettings) {
+          triggerSettings = value.autoCompleteTriggerSettings({
+            channel,
+            client,
+            emojiSearchIndex: value.emojiSearchIndex,
+            onMentionSelectItem: onSelectItem,
+          });
+        } else {
+          triggerSettings = ACITriggerSettings<StreamChatGenerics>({
+            channel,
+            client,
+            emojiSearchIndex: value.emojiSearchIndex,
+            onMentionSelectItem: onSelectItem,
+          });
+        }
       }
+      return triggerSettings;
+    } catch (error) {
+      console.warn('Error in getting trigger settings', error);
+      throw error;
     }
-    return triggerSettings;
   };
 
   const triggerSettings = getTriggerSettings();
