+++ package/src/utils/constants.ts
@@ -1,0 +1,6 @@
+export const defaultAutoCompleteSuggestionsLimit = 10;
+export const defaultMentionAllAppUsersQuery = {
+  filters: {},
+  options: {},
+  sort: {},
+};
